+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/AbstractBeanDefinitionReader.java
@@ -1,0 +1,34 @@
+package com.yrb.tinyioc;
+
+import com.yrb.tinyioc.io.ResourceLoader;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * @author bjyangrubing
+ * @createTime 2017/8/7 16:19
+ * Description: BeanDefinitionReader的抽象类
+ */
+public abstract class AbstractBeanDefinitionReader implements BeanDefinitionReader
+{
+	private Map<String, BeanDefinition> registry;
+
+	private ResourceLoader resourceLoader;
+
+	protected AbstractBeanDefinitionReader(ResourceLoader resourceLoader)
+	{
+		this.registry = new HashMap<String, BeanDefinition>();
+		this.resourceLoader = resourceLoader;
+	}
+
+	public Map<String, BeanDefinition> getRegistry()
+	{
+		return registry;
+	}
+
+	public ResourceLoader getResourceLoader()
+	{
+		return resourceLoader;
+	}
+}