+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/BeanDefinitionReader.java
@@ -1,0 +1,11 @@
+package com.yrb.tinyioc;
+
+/**
+ * @author bjyangrubing
+ * @createTime 2017/8/7 16:18
+ * Description: 从配置中读取BeanDefinition
+ */
+public interface BeanDefinitionReader
+{
+	void loadBeanDefinitions(String location) throws Exception;
+}